requirements.txt
pyyaml
datasets
evaluate
urllib3
transformers
tokenizers
torch 
torchvision 
torchaudio
tqdm
sentencepiece
accelerate
openai
